Separate routes and helper functions
